requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
proto-plus==1.13.0
protobuf==3.14.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pyspark==3.0.1
pyspark-stubs==3.0.0.post1
pytoml==0.1.21
pytz==2021.1
requests==2.25.1
retrying==1.3.3
rsa==4.7
six==1.15.0
urllib3==1.26.3
webencodings==0.5.1
